Fixed two bugs in evaluate_spatial_sharing_udp with sp1StartTime and sp2StartTime
The first bug was where the two values were not set until the simulation had begun, this meant that the Simulator schdule at the bottom of the script, which scduled the first MeasureOverSP1/ MeasureOverSP2 would always scdule them with variable 0.

The second bug fixed, is where the values for sp1StartTime and sp2StartTime were not actually set at thier respective start times, rather at the time that they ended, this caused problems when sending further channel quality requests.